fix: resolve rebase conflicts with main — exit code, LogSeverity cast, proto regen
- container_service.go: discard exit code from streamContainerOutput in
  StartContainer (v1 path; multi-service group branch added by main)
- telemetry_service_v2.go: cast *filter.MinSeverity (now LogSeverity enum) to
  int32 for comparison; required after min_severity changed from int32 to enum
- telemetry_service.pb.go: regenerate to fix corrupted raw descriptor bytes
  that resulted from manually merging conflict markers

+++ go/internal/agent/services/telemetry_service_v2.go
@@ -191,7 +191,7 @@ func filterLogsV2(req *otelpb.ExportLogsServiceRequest, filter *agentpbv2.Stream
 	appName := filter.AppName
 	var minSeverity int32
 	if filter.MinSeverity != nil {
-		minSeverity = *filter.MinSeverity
+		minSeverity = int32(*filter.MinSeverity)
 	}
 
 	if serviceName == nil && appName == nil && minSeverity == 0 {
